yolo-server/requirements.txt
# YOLO Detection Server Dependencies

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic-settings>=2.1.0
Pillow>=10.0.0
numpy>=1.24.0
ultralytics>=8.0.0

# OCR (ConvNextViT ONNX)
onnxruntime>=1.16.0
opencv-python-headless>=4.8.0
